Shrink generator/src/generator/namespace.rs
This commit moves everything but NamespaceGenerator to a sub-module.
That way, the file (which is currently out largest file by far) becomes
a lot smaller.
